Squash logic into MessageTypeToType
